Tout marche !!! aleluyah ! attention mode create activé
Tout marche !!! aleluyah ! attention mode create activé

J'ai du changer les modèles et passer l'id en int :s pour que ca marche avec les fonctions qu'on avait fait en cours .

+++ src/main/java/com/wha/controller/ConseillerRestController.java
@@ -29,7 +29,7 @@ public class ConseillerRestController {
 	}
 
 	@GetMapping("/conseillers/{id}")
-	public ResponseEntity<Conseiller> getConseiller(@PathVariable("id") Long id) {
+	public ResponseEntity<Conseiller> getConseiller(@PathVariable("id") int id) {
 
 		Conseiller conseiller = conseillerDao.get(id);
