+++ src/main/java/com/wha/model/Admin.java
@@ -42,7 +42,7 @@ public class Admin extends User {
 		this.contratStartingDate = contratStartingDate;
 	}
 
-	public Admin(Long id, String firstName, String lastName, String email, String mobile, String address,
+	public Admin(int id, String firstName, String lastName, String email, String mobile, String address,
 			Date dateOfBirth, String matricule, String fonction, Date contratStartingDate) {
 		super(id, firstName, lastName, email, mobile, address, dateOfBirth);
 		this.matricule = matricule;
