+++ src/main/java/com/wha/model/Conseiller.java
@@ -23,7 +23,7 @@ public class Conseiller extends User {
 		super();
 	}
 
-	public Conseiller(Long id, String firstName, String lastName, String email, String mobile, String address,
+	public Conseiller(int id, String firstName, String lastName, String email, String mobile, String address,
 			Date dateOfBirth, List<Client> clients) {
 		super(id, firstName, lastName, email, mobile, address, dateOfBirth);
 		this.clients = clients;
